+++ kinotic-domain/build.gradle
@@ -8,6 +8,7 @@ dependencies {
 
     implementation 'io.vertx:vertx-core'
     implementation 'io.vertx:vertx-auth-common'
+    implementation 'io.vertx:vertx-auth-jwt'
 
     implementation 'co.elastic.clients:elasticsearch-java'
     implementation 'co.elastic.clients:elasticsearch-rest5-client'
